Here is the commit message translated and written in English:
Implemented the upgrade of the session and auth services to return the UID for use in future functionality.
Implemented the restructuring of the entire application (performed a cleanup) to make it easier to add new pages and features, which will be more user-friendly.
Implemented the addition of a function that displays a JSON-type message on the screen, which in a future upgrade will be used to process data through stored procedures in the backend.